compilation warnings fixes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@20818 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/mac/carbon/toplevel.cpp
@@ -844,7 +844,7 @@ void wxTopLevelWindowMac::MacFireMouseEvent(
 
     if ( kind == mouseDown )
     {
-        if ( timestamp - gs_lastWhen <= GetDblTime() )
+        if ( timestamp - gs_lastWhen <= (long) GetDblTime() )
         {
             if ( abs( localwhere.h - gs_lastWhere.h ) < 3 && abs( localwhere.v - gs_lastWhere.v ) < 3 )
             {
